+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### `Added` for new features
+- class GenericRestApiJsonModel implements SeablastModelInterface
+- Exception\UploadException to translate (int) code to (string) message
 
 ### `Changed` for changes in existing functionality
 - improved SeablastMysqli error logging